pyproject.toml
[tool.poetry]
name = "t-test-shop"
version = "0.1.0"
description = ""
authors = ["Tryd0g0lik <work80@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
pytest-asyncio = "^0.24.0"

[tool.poetry.group.dev.dependencies]
asyncio = "^3.4.3"
flake8 = "^7.1.1"
pytest = "^8.3.3"
pre-commit = "^3.8.0"
black = "^24.8.0"
pre-commit-hooks = "^4.6.0"
pylint = "^3.2.7"
isort = "^5.13.2"
playwright = "^1.47.0"
pytest-playwright = "^0.5.2"
pytest-aio = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
